requirements.txt
boto3==1.28.65
cognitojwt==1.4.1
Flask==2.0.1
Flask_Cors==3.0.10
langchain==0.0.304
openai==0.27.8
python-dotenv==1.0.0
Werkzeug==2.2.2
faiss-cpu==1.7.4
